requirements.txt
asgiref==3.5.1
attrs==22.1.0
autopep8==1.7.0
boto3==1.26.74
botocore==1.29.74
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-mysql==4.6.0
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pika==1.2.1
Pillow==9.2.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.11
yapf==0.32.0
zipp==3.8.0
